Move getSpacing to scripts folder. Add height prop to Inline.
